backend/embedding_generator/requirements.txt
boto3>=1.34.0
opensearch-py>=2.4.0
python-dotenv>=1.0.0
